+++ src/api/publication/publication.controller.js
@@ -1,16 +1,17 @@
 const Publication = require('./publication.model')
-const Admin = require('../admin/admin.model')
+const User = require('../user/user.model')
 
 const uploadPublication = async (req, res) => {
   try {
     const publicationData = req.body
-    const id = req.admin
-    const admin = await Admin.findById(id)
-    const publication = await Publication.create({ ...publicationData, admin: id })
-    admin.news.push(publication)
-    await admin.save({ validateBeforeSave: false })
-
-    res.status(201).json({ message: 'Publication uploaded successfully', data: publication })
+    const id = req.user
+    const user = await User.findById(id)
+    if (user.isAdmin === true) { 
+      const publication = await Publication.create({ ...publicationData, user: id })
+      user.news.push(publication)
+      await user.save({ validateBeforeSave: false })
+      res.status(201).json({ message: 'Publication uploaded successfully', data: publication })
+    }
   } catch (err) {
     res.status(400).json({ message: 'Publication cannot be uploaded', data: err })
   }
@@ -32,23 +33,27 @@ const updatePublication = async (req, res) => {
   try {
     const  { publicationId } = req.params
     const updatedPublication = req.body
-    const id = req.admin
-    const admin = await Admin.findById(id)
-    const publicationUpdated = await Publication.findByIdAndUpdate(publicationId, updatedPublication, {new: true})
-    res.status(200).json({ message: 'Publication updated successfully', data: publicationUpdated})
+    const id = req.user
+    const user = await User.findById(id)
+    if (user.isAdmin === true) { 
+      const publicationUpdated = await Publication.findByIdAndUpdate(publicationId, updatedPublication, {new: true})
+      res.status(200).json({ message: 'Publication updated successfully', data: publicationUpdated})
+    } 
   } catch(err) {
-    res.status(404).json({ message: 'Publication could not be updated', data: err})
+    res.status(400).json({ message: 'Publication could not be updated', data: err})
   }
 }
 
 const destroyPublication = async (req, res) => {
   try {
     const { publicationId } = req.params
-    const id = req.admin
-    const admin = await Admin.findById(id)
-    const publication = await Publication.find({_id: publicationId, admin: id})
-    const deletedPublication = await Publication.deleteOne(product._id)
-    res.status(200).json({ message: 'Publication deleted', data: publication[0].title})
+    const id = req.user
+    const user = await User.findById(id)
+    if (user.isAdmin === true) { 
+      const publication = await Publication.find({_id: publicationId, user: id})
+      const deletedPublication = await Publication.deleteOne(product._id)
+      res.status(200).json({ message: 'Publication deleted', data: publication[0].title})
+    }
   } catch(err) {
     res.status(404).json({ message: 'Publication does not exist', data: err})
   }
